blog/requirements.txt
django
django_rest_framework
djangorestframework-simplejwt
environs
psycopg2-binary
